Add files via upload
In a school, the number of members in one class is not equal. As an honors student of programming, you are responsible for writing a program that receives the number of members of each group and the output number is the number that the members of each group must have, so that the number of people in the groups is equal. Note: The number of groups is uncertain and is not going to change. ## Guidance:

The end of the sequence of numbers or the number of members of the groups is determined by zero, which means that there are no more groups left. That is, you have to use loops in such a way that they close by taking zero input.
Since we don't know the number of inputs, we shouldn't define a variable based on their number. Please note that we only need their number and sum.
It is guaranteed that the sum of the members of the given groups is divisible by the number of groups.
Input 
An integer in each line
�
nIt will be given. The last entry is zero.
0
≤
�
≤
5
0
0≤n≤50

Output 
A number that indicates the number of members of the groups after their number is equal.